added query type rin ...
Query __in was backwards, it checked if the value was in the field, that
is functionality is now in new query __rin.

Query __in now checks if field is value so the logic matches __eq.

Strings make things complicated as they are iterable so special checks
are required to make sure both field and value are not strings when
using __in/__rin searches.