Query now always copies Manager._instances...
I was selectively copying the Manager._instances and being careful to
return a copy but guaranteed at some point in the future a mistake will
get made and data will be overwritten and people will be said. Most
likely me.

All for very questionable speed improvements. This will use more memory
however.

The "proper" thing to do (maybe/probably) would be to only operate on
Manager._instances.keys() and only return the values during
a Query.__get__ or __iter__.