big change, database now works with storage instances...
previously the database kept track of storage classes but now it stores
a dict of storage instances. This simplifies several things and arguably
it should have always been this way.

The big features this enables is locking the storage datafile as soon
as the storage object is created. This will prevent two instances
of the database (even in different processes) from opening the same file
and stomping on each others changes.

lots more tests to get back to 100% coverage.